+++ torch/csrc/jit/python/python_sugared_value.cpp
@@ -914,8 +914,11 @@ std::shared_ptr<SugaredValue> PythonExceptionValue::call(
             ->insertNode(caller.graph()->createTuple(message_values))
             ->output();
   }
+  Value* qualified_class_name =
+      insertConstant(*caller.graph(), qualified_name_, loc);
 
-  return std::make_shared<ExceptionMessageValue>(error_message);
+  return std::make_shared<ExceptionMessageValue>(
+      error_message, qualified_class_name);
 }
 
 bool isNamedTupleClass(const py::object& obj) {
