+++ torch/csrc/jit/python/python_sugared_value.h
@@ -328,7 +328,10 @@ struct VISIBILITY_HIDDEN PythonClassValue : public ClassValue {
 struct VISIBILITY_HIDDEN PythonExceptionValue : public ExceptionValue {
   explicit PythonExceptionValue(const py::object& exception_class)
       : ExceptionValue(
-            py::str(py::getattr(exception_class, "__name__", py::str("")))) {}
+            py::str(py::getattr(exception_class, "__name__", py::str("")))) {
+    qualified_name_ = py::str(py::module::import("torch._jit_internal")
+                                  .attr("_qualified_name")(exception_class));
+  }
 
   std::string kind() const override {
     return "Python exception";
@@ -340,6 +343,9 @@ struct VISIBILITY_HIDDEN PythonExceptionValue : public ExceptionValue {
       at::ArrayRef<NamedValue> args,
       at::ArrayRef<NamedValue> kwargs,
       size_t n_binders) override;
+
+ private:
+  std::string qualified_name_;
 };
 
 // Python Slice class.
