+++ torch/csrc/jit/runtime/jit_exception.cpp
@@ -5,5 +5,10 @@ namespace jit {
 
 JITException::JITException(const std::string& msg) : std::runtime_error(msg) {}
 
+CustomJITException::CustomJITException(
+    const std::string& msg,
+    const std::string& python_class_name)
+    : JITException(msg), python_class_name_(python_class_name) {}
+
 } // namespace jit
 } // namespace torch
